refactor: merge execution lift handling
