+++ packages/myst-cli/src/transforms/index.ts
@@ -9,5 +9,4 @@ export * from './include.js';
 export * from './links.js';
 export * from './mdast.js';
 export * from './outputs.js';
-export * from './inlineExpressions.js';
 export * from './types.js';
